Bug 971086 - [Camera] 1.4 Visual Design - Image Capture/ Video Buttons
