+++ apps/camera/js/app.js
@@ -7,6 +7,7 @@ define(function(require, exports, module) {
 
 var performanceTesting = require('performanceTesting');
 var ViewfinderView = require('views/viewfinder');
+var ControlsView2 = require('views/controls-2');
 var ControlsView = require('views/controls');
 var FocusRing = require('views/focus-ring');
 var lockscreen = require('lib/lock-screen');
@@ -115,9 +116,12 @@ App.prototype.runControllers = function() {
 };
 
 App.prototype.initializeViews = function() {
+  var newControls = this.settings.newControls.selected('value');
+  var Controls = newControls ? ControlsView2 : ControlsView;
+
   this.views.viewfinder = new ViewfinderView();
-  this.views.controls = new ControlsView();
   this.views.focusRing = new FocusRing();
+  this.views.controls = new Controls();
   this.views.hud = new HudView();
   debug('views initialized');
 };
