+++ apps/camera/js/controllers/controls.js
@@ -33,15 +33,16 @@ function ControlsController(app) {
 
 ControlsController.prototype.bindEvents = function() {
   this.app.settings.on('change:mode', this.controls.setter('mode'));
-  this.app.on('change:recording', this.controls.setter('recording'));
-  this.app.on('camera:timeupdate', this.controls.setVideoTimer);
-  this.controls.on('click:capture', this.app.firer('capture'));
-  this.controls.on('click:gallery', this.onGalleryButtonClick);
-  this.controls.on('click:switch', this.app.settings.mode.next);
-  this.controls.on('click:cancel', this.onCancelButtonClick);
-  this.app.on('camera:loading', this.disableButtons);
+  this.app.on('newimage', this.onNewMedia);
+  this.app.on('newvideo', this.onNewMedia);
   this.app.on('camera:ready', this.enableButtons);
   this.app.on('camera:busy', this.disableButtons);
+  this.app.on('change:recording', this.controls.setter('recording'));
+  this.app.on('camera:timeupdate', this.controls.setVideoTimer);
+  this.controls.on('tap:capture', this.app.firer('capture'));
+  this.controls.on('tap:gallery', this.onGalleryButtonClick);
+  this.controls.on('tap:switch', this.app.settings.mode.next);
+  this.controls.on('tap:cancel', this.onCancelButtonClick);
   debug('events bound');
 };
 
@@ -69,8 +70,11 @@ ControlsController.prototype.enableButtons = function() {
   this.controls.enable('buttons');
 };
 
-ControlsController.prototype.onSwitchClick = function() {
-  this.app.settings.get('mode').next();
+/**
+  When a new image is available it displays the thumbnail on the gallery button
+*/
+ControlsController.prototype.onNewMedia = function(image) {
+  this.controls.setThumbnail(image.thumbnail);
 };
 
 /**
