+++ apps/camera/js/lib/filmstrip.js
@@ -28,6 +28,7 @@ define(function(require, exports, module) {
   var constants = require('config/camera');
   var broadcast = require('lib/broadcast');
   var MediaFrame = require('MediaFrame');
+  var createThumbnailImage = require('lib/create-thumbnail-image');
 
   /**
    * Locals
@@ -45,11 +46,8 @@ define(function(require, exports, module) {
     var items = [];
     var currentItemIndex;
 
-    var DEVICE_RATIO = window.devicePixelRatio;
     // Maximum number of thumbnails in the filmstrip
     var MAX_THUMBNAILS = 5;
-    var THUMBNAIL_WIDTH = 46 * DEVICE_RATIO;  // size of each thumbnail
-    var THUMBNAIL_HEIGHT = 46 * DEVICE_RATIO;
 
     // Timer for auto-hiding the filmstrip
     var hideTimer = null;
@@ -63,9 +61,6 @@ define(function(require, exports, module) {
     var shareButton = document.getElementById('share-button');
     var deleteButton = document.getElementById('delete-button');
     debug('fetched elements');
-
-    // Offscreen image for generating thumbnails
-    var offscreenImage = new Image();
 
     // Set up event handlers
     cameraButton.onclick = hidePreview;
@@ -344,7 +339,7 @@ define(function(require, exports, module) {
           parseJPEGMetadata(previewBlob, function(thumbnailMetadata) {
             metadata.preview.width = thumbnailMetadata.width;
             metadata.preview.height = thumbnailMetadata.height;
-            createThumbnail(
+            createThumbnailImage(
               previewBlob,
               false,
               metadata.rotation,
@@ -370,7 +365,7 @@ define(function(require, exports, module) {
     function addVideo(video) {
       video.isVideo = true;
 
-      createThumbnail(
+      createThumbnailImage(
         video.poster.blob,
         true,
         video.rotation,
@@ -479,113 +474,6 @@ define(function(require, exports, module) {
         URL.revokeObjectURL(item.element.src);
       });
       items.length = 0;
-    }
-
-    // Create a thumbnail size canvas, copy the <img> or <video> into it
-    // cropping the edges as needed to make it fit, and then extract the
-    // thumbnail image as a blob and pass it to the callback.
-    function createThumbnail(imageBlob, video, rotation, mirrored, callback)
-    {
-      offscreenImage.src = URL.createObjectURL(imageBlob);
-      offscreenImage.onload = function() {
-        // Create a thumbnail image
-        var canvas = document.createElement('canvas');
-        var context = canvas.getContext('2d');
-        canvas.width = THUMBNAIL_WIDTH;
-        canvas.height = THUMBNAIL_HEIGHT;
-        var imgWidth = offscreenImage.width;
-        var imgHeight = offscreenImage.height;
-        var scalex = canvas.width / imgWidth;
-        var scaley = canvas.height / imgHeight;
-
-        // Take the larger of the two scales: we crop the image to the thumbnail
-        var scale = Math.max(scalex, scaley);
-
-        // Calculate the region of the image that will be copied to the
-        // canvas to create the thumbnail
-        var w = Math.round(THUMBNAIL_WIDTH / scale);
-        var h = Math.round(THUMBNAIL_HEIGHT / scale);
-        var x = Math.round((imgWidth - w) / 2);
-        var y = Math.round((imgHeight - h) / 2);
-
-        var centerX = Math.floor(THUMBNAIL_WIDTH / 2);
-        var centerY = Math.floor(THUMBNAIL_HEIGHT / 2);
-
-        // If a orientation is specified, rotate/mirroring the canvas context.
-        if (rotation || mirrored) {
-          context.save();
-          // All transformation are applied to the center of the thumbnail.
-          context.translate(centerX, centerY);
-        }
-
-        if (mirrored) {
-          context.scale(-1, 1);
-        }
-        if (rotation) {
-          switch (rotation) {
-          case 90:
-            context.rotate(Math.PI / 2);
-            break;
-          case 180:
-            context.rotate(Math.PI);
-            break;
-          case 270:
-            context.rotate(-Math.PI / 2);
-            break;
-          }
-        }
-
-        if (rotation || mirrored) {
-          context.translate(-centerX, -centerY);
-        }
-
-        // Draw that region of the image into the canvas, scaling it down
-        context.drawImage(offscreenImage, x, y, w, h,
-                          0, 0, THUMBNAIL_WIDTH, THUMBNAIL_HEIGHT);
-
-        // Restore the default rotation so the play arrow comes out correctly
-        if (rotation || mirrored) {
-          context.restore();
-        }
-
-        // We're done with the offscreen image now
-        URL.revokeObjectURL(offscreenImage.src);
-        offscreenImage.onload = null;
-        offscreenImage.src = '';
-
-        if (video) {
-          // Superimpose a translucent play button over
-          // the thumbnail to distinguish it from a still photo
-          // thumbnail. First draw a transparent gray circle.
-          context.fillStyle = 'rgba(0, 0, 0, .3)';
-          context.beginPath();
-          context.arc(THUMBNAIL_WIDTH / 2, THUMBNAIL_HEIGHT / 2,
-                               THUMBNAIL_HEIGHT / 3, 0, 2 * Math.PI, false);
-          context.fill();
-
-          // Now outline the circle in white
-          context.strokeStyle = 'rgba(255,255,255,.6)';
-          context.lineWidth = 2;
-          context.stroke();
-
-          // And add a white play arrow.
-          context.beginPath();
-          context.fillStyle = 'rgba(255,255,255,.6)';
-          // The height of an equilateral triangle is sqrt(3)/2 times the side
-          var side = THUMBNAIL_HEIGHT / 3;
-          var triangle_height = side * Math.sqrt(3) / 2;
-          context.moveTo(THUMBNAIL_WIDTH / 2 + triangle_height * 2 / 3,
-                                  THUMBNAIL_HEIGHT / 2);
-          context.lineTo(THUMBNAIL_WIDTH / 2 - triangle_height / 3,
-                                  THUMBNAIL_HEIGHT / 2 - side / 2);
-          context.lineTo(THUMBNAIL_WIDTH / 2 - triangle_height / 3,
-                                  THUMBNAIL_HEIGHT / 2 + side / 2);
-          context.closePath();
-          context.fill();
-        }
-
-        canvas.toBlob(callback, 'image/jpeg');
-      };
     }
 
     function setOrientation(orientation) {
