+++ apps/camera/style/main.css
@@ -15,6 +15,7 @@
 @import url("viewfinder.css");
 @import url("focusring.css");
 @import url("controls.css");
+@import url("controls-2.css");
 @import url("overlay.css");
 @import url("hud.css");
 @import url("filmstrip.css");
